docker, tests and environment details

+++ jest.globalSetup.ts
@@ -26,30 +26,32 @@ let myCustomOptions = {
 const port = process.env.SERVER_PORT
 
 export default async function setupServerToIntegrationTests() {
-   execSync('yarn prisma migrate reset -f')
+   if(!process.env.UNIT) {
+      execSync('yarn prisma migrate reset -f')
 
-   let serverToBeClosedAfter: any
-   const app = await returnApp()
+      let serverToBeClosedAfter: any
+      const app = await returnApp()
 
-   await new Promise(resolve => {
-      serverToBeClosedAfter = app.listen(port, () => {
-         // styling log
-         const message = `GlobalSetupFile: Server is running on http://localhost:${port}`
-         const styleTab = '-'.repeat(message.length +2)
-         const color = `\x1b[36m`
-         console.log(color,'\n' + `${styleTab}\n${styleTab}\n`,)
-         console.log(`\x1b[37m`, message)
-         console.log(color,`\n${styleTab}\n${styleTab}`)
-         // styling log
-         resolve('done')
-      }).on('error', (err) => {
-         console.error(err)
-         process.exit(1)
+      await new Promise(resolve => {
+         serverToBeClosedAfter = app.listen(port, () => {
+            // styling log
+            const message = `GlobalSetupFile: Server is running on http://localhost:${port}`
+            const styleTab = '-'.repeat(message.length +2)
+            const color = `\x1b[36m`
+            console.log(color,'\n' + `${styleTab}\n${styleTab}\n`,)
+            console.log(`\x1b[37m`, message)
+            console.log(color,`\n${styleTab}\n${styleTab}`)
+            // styling log
+            resolve('done')
+         }).on('error', (err) => {
+            console.error(err)
+            process.exit(1)
+         })
       })
-   })
 
-   Object.assign(globalThis, {
-      __testServer__: serverToBeClosedAfter,
-      __myCustomOptions__: myCustomOptions,
-   })
+      Object.assign(globalThis, {
+         __testServer__: serverToBeClosedAfter,
+         __myCustomOptions__: myCustomOptions,
+      })
+   }
 };
